Add files via upload
This is a Java project that focuses on I/O and can encipher, decipher, and break some of the common ciphers. This includes the ability to encode and decode Atbash, Shift Cipher, Repeating Keyword (Viginere), Playfair, and Affine. This program is not able to break the Playfair and Affine ciphers.